[broken build!] WIP to get coredata working properly
Halfway through setting up the coredata database and fetching the
appropriate record on selection of a row in the predictive search
results.

This is update highlights a change in design philosophy. I'll have to
decide by the next working build if we're creating a local version of
the contacts (seems the wrong way to go, but I'm following the example
from the 2011 Standofrd lecture on coredata). The better way to do it
(and what I wanted to do originally) would be to create a Person object
after the selection, and pass that along to the HomeViewController,
which would show a HUD with the Person's data to the user.