Passing Person data around view controllers
- no longer broken, but crashes after a while
- moved a number of things around:
   1) the context is no longer passed to controllers, there is a
coredata manager class that takes care of everything
   2) (possibly redundant) coredata manager now creates a Person object
when a new one is needed (c.f. w/ a) previous version, in which the
controller created it and b) a version in which the Person+Create class
would create one through a static method)
- Person data now gets transferred around the controllers
- a preliminary PersonDetailView has been stubbed in.

TODO:
- there seems to be a problem with the autosave of the context. Leave
the app running for a while and it crashes (yay) with the following
error message:
Terminating app due to uncaught exception
'NSInternalInconsistencyException', reason: 'This
NSPersistentStoreCoordinator has no persistent stores.  It cannot
perform a save operation.'

-Decide whether or not Person+Create is redundant. The DB Manager might
be enough.