integrated facebook
facebook works.
A note on the implementation of coredata: currently there are two
methodologies for saving data to local db, depending on the source:
a) addressbook - saves to coredata only when the pertinent people are
shown on screen by the predictive search
b) facebook - saves to coredata after initial call to facebook. MAJOR
ASSUMPTION: All friends are retrieved in this initial call. Need to
look into this, maybe fb splits up the results. The profile images are
only retrieved when the person they belong to gets shown on screen by
the predictive search. Once they are shown, they are saved to the local
db so this fetching an image from facebook nonsense is only done once
per person.

TODO:
- add debts to coredata
- fix those damn NSConstraints, they're breaking when the keyboard
leaves/reappears.
- fix the UI/UX (the whole point of this exercise)