basic aesthetic changes
changed some of the images around
there's a commented-out call to the db to fetch the Person records with
the most recent debts. This dataset should be displayed on the opening
page.

most urgent todo:
fix that call most recent persons call - screws up b/c of an unknown
issue relating to one-to-many relationship in coredata schema.